+++ flake.nix
@@ -1,0 +1,22 @@
+{
+  description = "MyClaw NixOS configuration for Hetzner Cloud";
+
+  inputs = {
+    nixpkgs.url = "github:NixOS/nixpkgs/nixos-24.11";
+    disko = {
+      url = "github:nix-community/disko";
+      inputs.nixpkgs.follows = "nixpkgs";
+    };
+  };
+
+  outputs = { self, nixpkgs, disko, ... }: {
+    nixosConfigurations.myclaw = nixpkgs.lib.nixosSystem {
+      system = "x86_64-linux";
+      modules = [
+        disko.nixosModules.disko
+        ./nixos/disk-config.nix
+        ./nixos/configuration.nix
+      ];
+    };
+  };
+}